Read spdlog and implement mouse event callbacks

+++ HideAndSeek/src/Engine/Event/MouseEvent.h
@@ -49,7 +49,7 @@ namespace Engine {
 	class ENGINE_API MouseScrolledEvent : public Event
 	{
 	public:
-		MouseScrolledEvent(float xOffset, float yOffset) 
+		MouseScrolledEvent(double xOffset, double yOffset) 
 			: m_XOffset(xOffset), m_YOffset(yOffset) {}
 
 		inline float GetXOffset() const { return m_XOffset; }
@@ -65,13 +65,13 @@ namespace Engine {
 		EVENT_CLASS_TYPE(MouseScrolled)
 
 	private:
-		float m_XOffset, m_YOffset;
+		double m_XOffset, m_YOffset;
 	};
 
 	class ENGINE_API MouseMovedEvent : public Event
 	{
 	public:
-		MouseMovedEvent(float x, float y)
+		MouseMovedEvent(double x, double y)
 			: m_MouseX(x), m_MouseY(y) {}
 
 		inline float GetX() const { return m_MouseX; }
@@ -86,7 +86,7 @@ namespace Engine {
 		EVENT_CLASS_CATEGORY(EventCategoryInput | EventCategoryMouse)
 		EVENT_CLASS_TYPE(MouseMoved)
 	private:
-		float m_MouseX, m_MouseY;
+		double m_MouseX, m_MouseY;
 	};
 
 }
